refactor(chat): consolidate 21 @deprecated aliases into chat/compat.ts (#1996)
Move the 21 historical `@deprecated` back-compat exports scattered across
`ag-ui.ts`, `types.ts`, `conversation.ts`, and `message-prep.ts` into a single
`src/chat/compat.ts` with a documented removal milestone.

Design avoids the circular import that reverted the earlier attempt (#2037):
`compat.ts` imports canonical symbols one-way from the public modules; those
modules never re-export from `compat.ts`. No cycle, so bundled SSR is unaffected
(verified: check:circular reports the same 2 pre-existing discovery/agent cycles
as main, none in chat).

- Add `./chat/compat` to the deno.json exports map so the deprecated symbols
  stay reachable for consumers mid-migration.
- Migrate the two test files that exercised the aliases to import them from
  `veryfront/chat/compat`.
- Regenerate the API reference (aliases now grouped under `veryfront/chat/compat`).

Closes #1996
